move define_heredoc_signals to signals.c rename redirect_file_name to get_label_name, refactor and finished handle_heredoc.c, handle expansions inside heredoc
