+++ includes/executes.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   executes.h                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: lalex-ku <lalex-ku@42sp.org.br>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/07 15:41:13 by lalex-ku          #+#    #+#             */
-/*   Updated: 2022/06/14 17:40:18 by sguilher         ###   ########.fr       */
+/*   Updated: 2022/06/15 14:49:25 by lalex-ku         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -46,7 +46,7 @@ char	*redirect_position(char *str, char redirect_char);
 char	next_redirect(char *str);
 void	close_all_fds(void);
 void	close_extra_fds(void);
-char	*redirect_file_name(char *redirect_str);
+char	*get_label_name(char *redirect_str);
 void	redirect_heredoc(char *command, int heredoc_number);
 
 // pipes
